Fix duplicate agent class locations
This commit resolves the issue reported in #20 where agent classes were defined in both:
- plangen/agents.py (as a module)
- plangen/agents/* (as a package)

Changes:
- Renamed agents.py to agents_legacy.py to clearly mark it as legacy
- Updated imports in plangen.py to use agents_legacy
- Updated __init__.py to properly import from agents_legacy
- Added deprecation warning to agents.py module
- Added README to the agents package explaining the structure
- Made agents.py re-export from agents_legacy.py for backward compatibility

These changes ensure that regardless of which import path is used, the same
agent implementations will be used, avoiding confusion and unexpected behavior.
