+++ plangen/__init__.py
@@ -7,14 +7,6 @@ This framework implements the PlanGEN workflow described in the paper
 
 __version__ = "0.1.0"
 
-# Import original classes for backward compatibility
-from .agents import (
-    ConstraintAgent,
-    SelectionAgent,
-    SolutionAgent,
-    VerificationAgent,
-)
-
 # Import public API classes - the recommended interface for users
 from .api import Algorithm, PlanGen, Verifiers, Visualization
 from .plangen import PlanGEN
@@ -22,6 +14,16 @@ from .visualization import (
     GraphRenderer,
     Observable,
     PlanObserver,
+)
+
+# Import original classes for backward compatibility
+# Using the module versions (not the package)
+# These are legacy interfaces and the implementation in plangen/agents.py
+from .agents_legacy import (
+    ConstraintAgent,
+    SelectionAgent,
+    SolutionAgent,
+    VerificationAgent,
 )
 
 __all__ = [
