+++ plangen/plangen.py
@@ -8,7 +8,7 @@ from langgraph.graph import END, StateGraph
 from langgraph.prebuilt import ToolNode
 from pydantic import BaseModel, Field
 
-from .agents import ConstraintAgent, SelectionAgent, SolutionAgent, VerificationAgent
+from .agents_legacy import ConstraintAgent, SelectionAgent, SolutionAgent, VerificationAgent
 from .models import BaseModelInterface, OpenAIModelInterface
 from .prompts import PromptManager
